requirements/dev.txt
-r base.txt

coverage==4.2
factory-boy==2.8.1
django-test-plus==1.0.16
